refactor: move mjs log util
