+++ src/svgo/log.mjs
